\name{MetaGenotypeCount}
\alias{MetaGenotypeCount}
\docType{data}
\title{
rs3131296 genetic association studies from SZGene database
}
\description{
rs3131296 genetic association studies from SZGene database.
}
\usage{data(MetaGenotypeCount)}
\format{
  The format is:
    22 lines: the first line is the names of columns; other lines are studies of rs3131296.
}
\source{
http://www.szgene.org/polydetail.asp?geneID=389&studyID=1587&ethnicDataID=3452
}
\references{
Stefansson H, Ophoff RA etc(Nature 2009 Aug 6;460(7256):744-7):Common variants conferring risk of schizophrenia.
}
\examples{
# data(MetaGenotypeCount)
}
\keyword{datasets}
