\name{MetaGenotypeData}
\alias{MetaGenotypeData}
\docType{data}
\title{
genotype raw data for rs3131296 genetic association studies
}
\description{
genotype raw data for rs3131296 genetic association studies, got by simulation based on the count of genotype from SZGene database.
}
\usage{data(MetaGenotypeData)}
\format{
  The format is:
   22 lines: the first line is the names of columns, other lines are information of studies.
	The 14-th and 15-th column separately are gentoype data for case and control samples. 
}
\references{
Stefansson H, Ophoff RA etc(Nature 2009 Aug 6;460(7256):744-7):Common variants conferring risk of schizophrenia.
}
\examples{
# data(MetaGenotypeData)
}
\keyword{datasets}
