\name{print.PermMeta}
\alias{print.PermMeta}
\title{
Print style for function 'meta.MCPerm' or 'meta.TradPerm'
}
\description{
Print style for function 'meta.MCPerm' or 'meta.TradPerm'.
}
\usage{
\method{print}{PermMeta}(x, \dots)
}
\arguments{
  \item{x}{
the result of function 'meta.MCPerm' or 'meta.TradPerm'.
}
 \item{\dots}{other arguments.}
}
\author{
Lanying Zhang and Yongshuai Jiang <jiangyongshuai@gmail.com>
}
\seealso{
  \code{\link{meta.MCPerm}},
  \code{\link{meta.TradPerm}}
}
\keyword{ PermMeta }
